refactor: Moved base class into seperate headers
