+++ rmw_cyclonedds_cpp/src/BaseCDRWriter.hpp
@@ -1,0 +1,45 @@
+// Copyright 2019 Rover Robotics via Dan Rose
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#ifndef BASE_CDR_WRITER_HPP_
+#define BASE_CDR_WRITER_HPP_
+
+#include <cstddef>
+#include <memory>
+
+#include "TypeSupport2.hpp"
+
+namespace rmw_cyclonedds_cpp
+{
+
+class BaseCDRWriter
+{
+public:
+  virtual size_t get_serialized_size(const void * data, SampleOrKey what) const = 0;
+  virtual size_t get_serialized_size_estimate(const void * data, SampleOrKey what) const = 0;
+
+  // includes 4 bytes encoding header
+  virtual size_t get_min_serialized_size(SampleOrKey what) const = 0;
+  // includes 4 bytes encoding header, SIZE_MAX if unbounded
+  virtual size_t get_max_serialized_size(SampleOrKey what) const = 0;
+
+  virtual void serialize(void * dest, const void * data, SampleOrKey what) const = 0;
+  virtual TypeGenerator type_generator() const = 0;
+  virtual ~BaseCDRWriter() = default;
+};
+
+std::unique_ptr<BaseCDRWriter> make_cdr_writer(
+  MessageMembersVariant members, SampleOrRequest variant);
+
+}  // namespace rmw_cyclonedds_cpp
+#endif  // BASE_CDR_WRITER_HPP_
