+++ rmw_cyclonedds_cpp/src/Serialization.hpp
@@ -17,6 +17,8 @@
 #include <memory>
 #include <vector>
 
+#include "BaseCDRWriter.hpp"
+#include "BaseCDRReader.hpp"
 #include "serdata.hpp"
 
 namespace rmw_cyclonedds_cpp
@@ -25,60 +27,6 @@ namespace rmw_cyclonedds_cpp
 std::pair<MessageMembersVariant, MessageMembersVariant>
 make_request_response_value_types(const rosidl_service_type_support_t * svc);
 
-enum class SampleOrKey
-{
-  Sample,
-  Key
-};
-
-enum class SampleOrRequest
-{
-  Sample,
-  Request
-};
-
-class BaseCDRWriter
-{
-public:
-  virtual size_t get_serialized_size(const void * data, SampleOrKey what) const = 0;
-  virtual size_t get_serialized_size_estimate(const void * data, SampleOrKey what) const = 0;
-
-  // includes 4 bytes encoding header
-  virtual size_t get_min_serialized_size(SampleOrKey what) const = 0;
-  // includes 4 bytes encoding header, SIZE_MAX if unbounded
-  virtual size_t get_max_serialized_size(SampleOrKey what) const = 0;
-
-  virtual void serialize(void * dest, const void * data, SampleOrKey what) const = 0;
-  virtual TypeGenerator type_generator() const = 0;
-  virtual ~BaseCDRWriter() = default;
-};
-
-std::unique_ptr<BaseCDRWriter> make_cdr_writer(
-  MessageMembersVariant members,
-  SampleOrRequest variant);
-
-class BaseCDRReader
-{
-public:
-  virtual void deserialize(
-    void * dest, const void * cdr, size_t cdrsize,
-    SampleOrKey what) const = 0;
-  virtual void extractkey(
-    std::vector<byte> & dest, const void * cdr, size_t cdrsize,
-    SampleOrKey what) const = 0;
-  virtual void extractkey_be(
-    std::vector<byte> & dst, const void * cdr, size_t cdrsize,
-    SampleOrKey what) const = 0;
-  virtual size_t print(
-    char * dst, size_t dstsize, const void * cdr, size_t cdrsize,
-    SampleOrKey what) const =  0;
-
-  virtual ~BaseCDRReader() = default;
-};
-
-std::unique_ptr<BaseCDRReader> make_cdr_reader(
-  MessageMembersVariant members,
-  SampleOrRequest variant);
 }  // namespace rmw_cyclonedds_cpp
 
 #endif  // SERIALIZATION_HPP_
